fix up networking if we started with DHCP

+++ src/karmalb-pkg/files/opt/klb/app/zbin/zlb-firstboot.sh
@@ -38,5 +38,25 @@ sed -i 's/^\# Required-Stop:.*/# Required-Stop:\t\$network \$remote_fs \$syslog 
 insserv -r snmpd
 
 
+# add an interface file if we were installed with DHCP etc.
+CDIR=/opt/klb/config
+NCONF=/etc/network/interfaces
+
+if ! grep -qi "inet static" $NCONF; then
+	EG="`ip route list|awk '/default/ { print $5 " " $3; exit }'`"
+	IF="`echo $EG|awk '{ print $1 }'`"
+	GW="`echo $EG|awk '{ print $2 }'`"
+	IPM="`/sbin/ifconfig $IF | awk '/inet / { print $2 " " $4 }'`"
+	IP="`echo $IPM| sed 's/addr:\([^ ]*\).*/\1/'`"
+	MASK="`echo \"$IPM\" | sed 's/.*Mask://'`"
+	echo "$IF::$IP:$MASK:up::" > $CDIR/if_${IF}_conf
+	sed -i "s/defaultgw=.*/defaultgw=$GW" $CDIR/global.conf
+	sed -i "s/defaultgwif=.*/defaultgwif=$IF" $CDIR/global.conf
+	echo "#zenmodified" > $NCONF
+	echo "auto lo" >> $NCONF
+	echo "iface lo inet loopback" >> $NCONF
+fi
+
+
 #
 rm $fb
